Refactor endpoint rule loader helpers
